changed saved PRICE value to integer and fixed import name

+++ dissen/signup/models/package.py
@@ -1,9 +1,9 @@
-from django import models
+from django.db import models
 
 class Package(models.Model):
     """Describes a package."""
 
-    PRICE = models.DecimalField(max_digits = 9, decimal_places = 2)
+    PRICE = models.IntegerField()
     DESCRIPTION = models.CharField(max_length = 75)
     TIMESTAMP = models.DateTimeField(auto_now_add = True)
     # todo: add reference to a company
